// src/app/types/filteredParticipant.ts
export interface filteredParticipant {
  "#": number;
  "Apellido paterno": string;
  "Apellido materno": string;
  "Nombre": string;
  "Prueba": string;
  "Edad": number;
  "TBW": number;
  "Agarre": number;
  "Puntos": number;
  "Salto": number;
  "Puntos_1": number;
  "Agilidad": number;
  "Puntos_2": number;
  "Resistencia": string;
  "Puntos_3": number;
  "Total": number;
}